job-manager: add jobtap plugin to bypass allocation submit to flux-kube
Add a jobtap plugin based on alloc-bypass.so. The plugin looks
for attributes.system.flux-kube set in the jobspec. If it's
present, the plugin sets the "alloc-bypass" flag, passes the jobspec
to flux-kube.py via RPC, and unpacks the response. Then the plugin
generates a dummy R (since the resources will reside on
OpenShift) and commits it to the KVS. Finally, the
plugin emits an "alloc" event.