update scheduler Dockerfile to use flux-sched
Problem: the current src/build/scheduler/Dockerfile is using an older
flux-sched branch, and flux-sched has since been refactored to build
with cmake.
Solution: update the Dockerfile to use the latest flux-sched,
and find libraries based on cmake install. Note that several of
the needed libs are currently built as STATIC and need to be
SHARED for this to work. We are currently patching with sed
and will improve upon this if/when a solution is added upstream.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>