update fluence scheduler build to use newer API
Problem: we made changes to the flux-sched reapi go module, and did not update here.
Solution: update the client interface here, ensuring that we also install exactly
the version of flux-sched that we are cloning. I am also cleaning up the README to
not use $ before commands (easier to copy paste), the go.mod/sum, and the utils
function to print an error code and not an integer (part of the refactor)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>