ci: add automated and on demand testing of fluence
Problem: we cannot tell if/when fluence builds will break against upstream
Solution: have a weekly run that will build and test images, and deploy on
successful results. For testing, I have added a complete example that uses
Job for fluence/default-scheduler, and the reason is because we can run
a container that generates output, have it complete, and there is no crash
loop backoff or similar. I have added a complete testing setup using kind,
and it is in one GitHub job so we can build both containers and load into
kind, and then run the tests. Note that MiniKube does NOT appear to work
for custom schedulers - I suspect there are extensions/plugins that need
to be added. Finally, I was able to figure out how to programmatically
check both the pod metadata for the scheduler along with events, and
that combined with the output should be sufficient (for now) to test
that fluence is working.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>