wip: experimental work on fluence
I am testing adding the abstraction of a pod group
to be carried from fluence labels (on the pods) through
to the same memory cache we are currently using for the
list of pods. As long as store the timestamp there (and
the cache is created once) I think this might be a good
replacement, but I also might have just broke everything
and it is a terrible idea.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>