update: pods seem to be scheduling in groups
Testing with a large cluster, I do not see interleaving as before, and I have changed
a lot so I am not sure what (might have) resolved that. I say might have because
I am not sure looking at the logs is the best means to determine this. For this
set of changes, I realized that for smaller groups created later, using a time.Now(I)
timestamp did not make sense - we would do better at least using the pod creation time
stamp. Either way, we always take the earlier one. I also realized that kubernetes
uses a metav1.Time, which wraps time.Time so I want to / changed it to use that
instead. Right now I am not cleaning up the pod group cache anywhere because I am
not sure where that place it. It should be OK to run one experiment set and get
a sense of fluence vs. default. Also, I realize that we have two different
execution pathways for pods assigned a group and those not. I am trying an
approach that creates a group for every actual group and pod without a
group (a group with one member) so the AskFlux and other associated logic is
handled the same. E.g., we do not have to try getting a group and then
check for a name or similar - there is always a group.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>