fix: use podgroup millisecond precision timestamp
Problem: the podgroups with second precision have interleaving
Solution: try to create an internal representation with better
precision. This looks promising with early testing, but I need
to consider the edge cases and how to clean up the groups, otherwise
a pod group might be re-created later and still in the cache.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>