feat: add start of webhook
Problem: we want to try a design where a mutating admission webhook can handle
receiving and creating PodGroup from labels. We are choosing mutating with
the expectation that, at some point, we might be able to change the size
(min/max/desired) either for the PodGroup or some other watcher to jobs.
Note that this is an empty skeleton - the webhook is added and running
but basically doing nothing. I am also fixing a bug that I noticed while
running kind that fluence was assigning work to the control plane. I think
there maybe used to be logic (a commented out worker label) that was
anticipating doing a check for a control plane, but it looks like on
production clusters we do not always haave access and it was never
finished. Note that this addition does not guarantee this design will
work, but it is just one step. Since the helm charts are manually
genereated for the scheduler-plugin (as far as I can tell) this
took me almost 6 hours to figure out and get working. I am really
starting to think there is no skill behind software engineering
beyond absolute patience.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>