pod-group: labels for name and size now lead to creation
Problem: we want the labels (size and name) that are explicitly set
to lead to the creation of the pod group so the user does not need to.
This is done by way of a watcher on pod, which will trigger after
the webhook that ensures that every pod (in a job or single pod)
has the proper label. Likely we want to do this for other abstractions
that hold pods as well, because it ensures that no matter how the pods
go into pending, we have the correct size and name. The only case that
a pod can come in without the label means that it was not scheduled
by fluence. The user is directed to not do this, but it is not
impossible (e.g., fluence sees itself show up here actually). So after
this addition we have the full steps to add the labels and create
the pod group, and next steps are (finally) to integrate this into
fluence (and remove the old abstraction to store it in memory).

Signed-off-by: vsoch <vsoch@users.noreply.github.com>