feat: add support for other abstractions
Problem: we need to be able to run deployments, stateful/replica sets
and have them handled by fluence.
Solution: allow the webhook to create pod groups for them. In the case
they are not targeted for fluence (any abstraction) and get into the
PreFilter, allow creation of a FauxPodGroup that will simply schedule one
job for the pod. We do this twice - in PreFilter and in the events
for update/delete.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>