bug: the metav1.MicroTime was not being set
Problem: I noticed in testing that the time only had granularity down to the second.
Solution: It appears that when we do a create of the PodGroup from the reconciler watch,
the metadata (beyond name and namespace) does not stick. I am not sure why, but the labels
are still retrievable from the pods (via the mutating webhook) after. So instead, we need
to get the size and creation timestamp at the first hit in reconcile, which (given how that
works) should still somewhat honor the order. I did try adding the timestamp to a label
but it got hairy really quickly (kept me up about 3 hours longer than I intended to!) The
good news now is that I see the microseconds in the Schedule Start Time, so we should be
almost ready to test this on a GCP cluster. I also had lots of time waiting for the containers
to rebuild so I made a diagram of how it is currently working. I have some concerns about
the internal state of fluxion (my kind cluster stopped working after some hours and I do not
know why) but we can address them later. We mostly need to see if there are jobs that are being
forgotten, etc.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>