go: update to 1.21
Problem: we need to update to a newer go to keep up
with the sig-scheduler upstream, and also the rainbow
scheduler integration.
Solution: upgrade to 1.21. This also required some
refactor of the main.go and fluence due to changes
in function signatures. This is a test to see if
tests are passing - the fluxion-go bindings used
here are from a branch (not merged yet) that can
be used for the PR this one is going into, and
before merging that final one we should merge
and release the bindings more properly.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>