chore: rename InitJGF to NewFluxJGF
Problem: The convention for creating new structs is typically
to use "New$Name" and return a reference or copy to it.
Solution: Rename InitJGF to NewFluxJGF to mirror that
practice, making it clear we get back a "new flux JGF"
and the function is not initializing something that
already exists.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>