jgf: refactor to use shared functions and fix containment
Problem: the containment paths are currently not set,
and each resource type has a separate function.
Solution: create a shared node generation function
that uses a common resource counter, where the counter
manages the global and resource-specific counts. This
scoped commit includes refactor of those JGF functions,
which also means removing resources that are not present
in the graph (e.g., rack and socket) and ensuring that
we do not recreate the subnet since we loop through
nodes (and might hit the same one twice). I am also
removing the NFD features because I think they are
changed and might lead to error if someone has them.
We should assess which of them we want/need and add
them back strategically. I have a local test but want
to push this first to give a look over / run current
tests to determine if other changes are needed.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>