test: add start of tests for fluence package
Problem: we will want to have scoped tests for fluence.
Solution: to start, add a simple test file for JGF (the
content of this current PR) and have an easy way to run
it with "make test" from that directory. Currently this
does not validate most of the structure - I would like
to review the output, discuss a testing strategy, and then
update the commit here to reflect that decision.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>