reorg: make build is working for kube-scheduler
This strategy refactors the fluence organization to build an entirely
in-tree plugin, having the fluence logic go under pkg/scheduler/framework
plugins, and then the sidecar container still serving the fluxion service.
The webhook and PodGroup as a controller are removed for now, because
either I need to add them back differently or remove all together. I have
extracted the charts from sig-scheduler-plugins and removed this
repository entirely as a dependency. This build currently works
to build the kube-scheduler (the new Dockerfile in the root) and
next I need to think about the actual design. The complexity of
getting the build environment / structure setup and working, period,
should not be overlooked! I suspect this is why they have made
the scheduler-plugins repository to begin with. Either way, having
fun :)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>