fix: upstream changes for build and charts
Problem: the current fluence build is broken.
Solution: the issue is two fold, and unfortunately depends on the
docker version. For newer docker that requires --load, that argument
would be needed in the build script, and we will need to add it (and
then add the script to our assets) if a newer version is used in the
CI. The other breaking change was changing the location of the kube-
scheduler binary in the controller image. Both of these should
now be fixed, and I will update the PR further pending results
in CI.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>