bug: fluxion no longer has "name" as a jgf field
Problem: name was removed (and replaced with Subsystem) and this is problematic
because the field comes through as nil and triggers an error. We do not
currently have subsystem support in fluence, and so I am opting to remove it
entirely for the simplest solution.
Solution: remove the Name field, which was not used aside from a print statement.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>