build: add custom scheduler Dockerfile
Problem: the sig-scheduler-plugins Dockerfile was updated to use
a distroless image over alpine, and this destroyed our logging
strategy so it was impossible to see printed messages from fluence.
Since this has bit us before, it will be a simpler strategy moving
forward to put this build under our control, and resort back to using
the previous alpine base.
Solution: add the src/build/scheduler/Dockerfile to our maintained
files, and ensure we use an alpine base that has a filesystem.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>